+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fix: Change overflow default value to manage negative margins
   on `k-ProjectCard`.
+- Feature: Add `MenuIcon` and component.
+- Feature: Add `KissKissBankBankLogoIcon` component.
+- Feature: Add `InformationIcon` component.
 
 ## [11.2.0] - 2017-04-26